Moves files into place and adds a simple error handle
