+++ README.md
@@ -1,3 +1,5 @@
+# Palette-Picker
+
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
 ## Available Scripts
