+++ src/Containers/App/App.test.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+
+describe('App', () => {
+  describe('app load', () => {
+    it('should match snapshot', () => {
+  
+    })
+    it('should have default state', () => {
+  
+    })
+  })
+  describe('mapStateToProps', () => {
+
+  })
+  describe('mapDispatchToProps', () => {
+    
+  })
+})
+
+
+// it('renders without crashing', () => {
+//   const div = document.createElement('div');
+//   ReactDOM.render(<App />, div);
+//   ReactDOM.unmountComponentAtNode(div);
+// });
