+++ src/Containers/App/_app.scss
@@ -1,3 +1,11 @@
+.App {
+  text-align: center;
+}
+
+h1 {
+  color: red;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
