+++ src/Style/index.scss
@@ -1,0 +1,4 @@
+@import './normalize.css';
+@import './variables';
+@import './mixins';
+@import '../Containers/App/app'
