+++ src/index.js
@@ -1,11 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import {BrowserRouter } from 'react-router-dom'
+import { BrowserRouter } from 'react-router-dom'
 import { Provider } from "react-redux";
 import { rootReducer } from "./Reducers";
 import { createStore } from "redux";
-import "./index.css";
-import App from "./App";
+import "./Style/index.scss";
+import App from "./Containers/App/App";
 import * as serviceWorker from "./serviceWorker";
 
 const store = createStore(
@@ -22,7 +22,4 @@ ReactDOM.render(
   document.getElementById("root")
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
 serviceWorker.unregister();
